refactor: move text preview loading from onRef into data pipeline
Parse body render blocks in allMessages$ computed and enrich with
text preview computeds. Replace render-driven onRef side effect with
useLastResolved consumption in TextPreview. Extract shared types and
helpers into parse-body-blocks.ts. Remove .then() from fetchPreviewText.
